// src/util/interfaces/gametypes.ts
export type BedwarsTypes =
  | ""
  | "eight_one_"
  | "eight_two_"
  | "four_three_"
  | "four_four_"
  | "two_four_";

export type ArcadeTypes =
  | "dayone"
  | "oneinthequiver"
  | "dragonwars2"
  | "farm_hunt"
  | "enderspleef"
  | "soccer"
  | "sw"
  | "hole_in_the_wall"
  | "hide_and_seek"
  | "simon_says"
  | "mini_walls"
  | "party"
  | "scuba_simulator"
  | "throw_out"
  | "zombies";

export type CrazyWallsTypes = "solo" | "solo_chaos" | "team" | "team_chaos";

export type MMTypes =
  | ""
  | "_MURDER_CLASSIC"
  | "_MURDER_HARDCORE"
  | "_MURDER_ASSASSINS";

export type SkyClashTypes = "" | "_solo" | "_doubles" | "_team_war";

export type SkyWarsTypes =
  | ""
  | "_solo_normal"
  | "_solo_insane"
  | "_team_normal"
  | "_team_insane"
  | "_ranked";

export type SmashTypes = "" | "_normal" | "_2v2" | "_teams";